getting wikipedia working for the most part

+++ src/main/java/com/oltpbenchmark/benchmarks/wikipedia/procedures/GetPageAnonymous.java
@@ -109,8 +109,8 @@ public class GetPageAnonymous extends Procedure {
             throw new UserAbortException(msg);
         }
 
-        int revisionId = rs.getInt("rev_id");
-        int textId = rs.getInt("rev_text_id");
+        long revisionId = rs.getLong("rev_id");
+        long textId = rs.getLong("rev_text_id");
         assert !rs.next();
         rs.close();
 
@@ -120,7 +120,7 @@ public class GetPageAnonymous extends Procedure {
         // "SELECT old_text,old_flags FROM `text` WHERE old_id = '"+textId+"' AND old_page = '"+pageId+"' LIMIT 1";
         // For now we run the original one, which works on the data we have
         st = this.getPreparedStatement(conn, selectText);
-        st.setInt(1, textId);
+        st.setLong(1, textId);
         rs = st.executeQuery();
         if (!rs.next()) {
             String msg = "No such text: " + textId + " for page_id:" + pageId + " page_namespace: " + pageNamespace + " page_title:" + pageTitle;
