+++ src/main/java/com/oltpbenchmark/benchmarks/wikipedia/procedures/UpdatePage.java
@@ -128,10 +128,10 @@ public class UpdatePage extends Procedure {
     // RUN
     // -----------------------------------------------------------------
 
-    public void run(Connection conn, int textId, int pageId,
+    public void run(Connection conn, long textId, int pageId,
                     String pageTitle, String pageText, int pageNamespace,
                     int userId, String userIp, String userText,
-                    int revisionId, String revComment, int revMinorEdit) throws SQLException {
+                    long revisionId, String revComment, int revMinorEdit) throws SQLException {
 
         boolean adv;
         PreparedStatement ps = null;
@@ -151,7 +151,7 @@ public class UpdatePage extends Procedure {
         rs = ps.getGeneratedKeys();
         adv = rs.next();
         assert (adv) : "Problem inserting new tuples in table text";
-        int nextTextId = rs.getInt(1);
+        long nextTextId = rs.getLong(1);
         rs.close();
         assert (nextTextId >= 0) : "Invalid nextTextId (" + nextTextId + ")";
 
@@ -159,7 +159,7 @@ public class UpdatePage extends Procedure {
         ps = this.getPreparedStatementReturnKeys(conn, insertRevision, new int[]{1});
         param = 1;
         ps.setInt(param++, pageId);       // rev_page
-        ps.setInt(param++, nextTextId);   // rev_text_id
+        ps.setLong(param++, nextTextId);   // rev_text_id
         ps.setString(param++, revComment);// rev_comment
         ps.setInt(param++, revMinorEdit); // rev_minor_edit // this is an error
         ps.setInt(param++, userId);       // rev_user
@@ -167,13 +167,13 @@ public class UpdatePage extends Procedure {
         ps.setString(param++, timestamp); // rev_timestamp
         ps.setInt(param++, 0);            // rev_deleted //this is an error
         ps.setInt(param++, pageText.length()); // rev_len
-        ps.setInt(param++, revisionId);   // rev_parent_id // this is an error
+        ps.setLong(param++, revisionId);   // rev_parent_id // this is an error
 //	    ps.execute();
         execute(conn, ps);
 
         rs = ps.getGeneratedKeys();
         adv = rs.next();
-        int nextRevId = rs.getInt(1);
+        long nextRevId = rs.getLong(1);
         rs.close();
         assert (nextRevId >= 0) : "Invalid nextRevID (" + nextRevId + ")";
 
@@ -182,7 +182,7 @@ public class UpdatePage extends Procedure {
         // anyway
         ps = this.getPreparedStatement(conn, updatePage);
         param = 1;
-        ps.setInt(param++, nextRevId);
+        ps.setLong(param++, nextRevId);
         ps.setString(param++, timestamp);
         ps.setInt(param++, pageText.length());
         ps.setInt(param++, pageId);
@@ -206,8 +206,8 @@ public class UpdatePage extends Procedure {
         ps.setInt(param++, userId);           // rc_user
         ps.setString(param++, userText);      // rc_user_text
         ps.setString(param++, revComment);    // rc_comment
-        ps.setInt(param++, nextTextId);       // rc_this_oldid
-        ps.setInt(param++, textId);           // rc_last_oldid
+        ps.setLong(param++, nextTextId);       // rc_this_oldid
+        ps.setLong(param++, textId);           // rc_last_oldid
         ps.setInt(param++, 0);                // rc_bot
         ps.setInt(param++, 0);                // rc_moved_to_ns
         ps.setString(param++, "");            // rc_moved_to_title
